Move time cluster into module
