+++ lib/ahoy_activity_utils.rb
@@ -31,7 +31,7 @@ module AhoyActivityUtils
 
                     timestamps = event.map{|e| e[:time] }
                     
-                    minutes = UserCollectionTime.total_contiguous_seconds(timestamps) / 60
+                    minutes = self.total_contiguous_seconds(timestamps) / 60
                     
                     unless minutes <= 0
                         begin 
@@ -51,5 +51,17 @@ module AhoyActivityUtils
                 end
             end
         end
+    end
+    def self.total_contiguous_seconds(times, tolerance=60.minutes)
+        
+        total_seconds = 0
+        from_time = nil
+
+        for time in times.sort do
+            time_diff = from_time.nil? ? 0 : (time - from_time).round
+            total_seconds += time_diff if time_diff < tolerance
+            from_time = time
+        end
+        return total_seconds
     end
 end
