+++ spec/models/ahoy_activity_utils_spec.rb
@@ -1,13 +1,14 @@
 require 'spec_helper'
+require 'ahoy_activity_utils'
 
-RSpec.describe UserCollectionTime do
+RSpec.describe AhoyActivityUtils do
 
   describe 'Static Methods' do
     it "finds correct difference between two times" do
         after   = 15.minutes.ago
         before  = 20.minutes.ago
 
-        duration = UserCollectionTime.total_contiguous_seconds([before, after])
+        duration = AhoyActivityUtils.total_contiguous_seconds([before, after])
 
         expect(duration).to eq(5.minutes)
     end
@@ -16,7 +17,7 @@ RSpec.describe UserCollectionTime do
         after   = 15.minutes.ago
         before  = 20.minutes.ago
 
-        duration = UserCollectionTime.total_contiguous_seconds([before, after, later])
+        duration = AhoyActivityUtils.total_contiguous_seconds([before, after, later])
 
         expect(duration).to eq(10.minutes)
     end
@@ -25,9 +26,9 @@ RSpec.describe UserCollectionTime do
         after   = 15.minutes.ago
         before  = 20.minutes.ago
 
-        duration_a = UserCollectionTime.total_contiguous_seconds([before, after, later])
-        duration_b = UserCollectionTime.total_contiguous_seconds([after, later, before])
-        duration_c = UserCollectionTime.total_contiguous_seconds([later, before, after])
+        duration_a = AhoyActivityUtils.total_contiguous_seconds([before, after, later])
+        duration_b = AhoyActivityUtils.total_contiguous_seconds([after, later, before])
+        duration_c = AhoyActivityUtils.total_contiguous_seconds([later, before, after])
 
         expect(duration_a).to eq(10.minutes)
         expect(duration_b).to eq(10.minutes)
@@ -38,7 +39,7 @@ RSpec.describe UserCollectionTime do
         after   = 10.minutes.ago
         before  = 70.minutes.ago
 
-        duration = UserCollectionTime.total_contiguous_seconds([before, after])
+        duration = AhoyActivityUtils.total_contiguous_seconds([before, after])
 
         expect(duration).to eq(0)
     end
@@ -47,7 +48,7 @@ RSpec.describe UserCollectionTime do
         after   = 10.minutes.ago
         before  = 70.minutes.ago
 
-        duration = UserCollectionTime.total_contiguous_seconds([before, after], 61.minutes)
+        duration = AhoyActivityUtils.total_contiguous_seconds([before, after], 61.minutes)
 
         expect(duration).to eq(60.minutes)
     end
@@ -55,14 +56,14 @@ RSpec.describe UserCollectionTime do
     it "sums contiguous elements, skips non-contiguous gaps" do
         times = [1.day.ago, (1.day.ago + 30.minutes), 60.minutes.ago, 30.minutes.ago]
         
-        duration = UserCollectionTime.total_contiguous_seconds(times)
+        duration = AhoyActivityUtils.total_contiguous_seconds(times)
         
         expect(duration).to eq(60.minutes)
     end
     it "sums shuffled timestamps correctly" do
         times = [1.day.ago, (1.day.ago + 30.minutes), 60.minutes.ago, 30.minutes.ago].shuffle
         
-        duration = UserCollectionTime.total_contiguous_seconds(times)
+        duration = AhoyActivityUtils.total_contiguous_seconds(times)
         
         expect(duration).to eq(60.minutes)
     end
